[temp] This is just a copy of LexMixPolly attempted to be broken up
What this does is attempt to unwire the coupling in lexMixPolly which
was to mixed success.  Ultimately it was determined that we need android
for lex there is no way around it best I can tell.  So this in itself
isn't used in the main project and was a falst attempt here for review
only.